+++ editor/src/components/canvas/canvas-strategies/strategies/grid-move-rearrange-strategy.ts
@@ -1,6 +1,17 @@
+import type { ElementPath } from 'utopia-shared/src/types'
 import { MetadataUtils } from '../../../../core/model/element-metadata-utils'
 import * as EP from '../../../../core/shared/element-path'
-import { isInfinityRectangle } from '../../../../core/shared/math-utils'
+import type {
+  ElementInstanceMetadata,
+  ElementInstanceMetadataMap,
+  GridContainerProperties,
+} from '../../../../core/shared/element-template'
+import { gridPositionValue } from '../../../../core/shared/element-template'
+import { isInfinityRectangle, offsetPoint } from '../../../../core/shared/math-utils'
+import { absolute } from '../../../../utils/utils'
+import type { CanvasCommand } from '../../commands/commands'
+import { reorderElement } from '../../commands/reorder-element-command'
+import { showGridControls } from '../../commands/show-grid-controls-command'
 import { controlsForGridPlaceholders } from '../../controls/grid-controls-for-strategies'
 import type { CanvasStrategyFactory } from '../canvas-strategies'
 import { onlyFitWhenDraggingThisControl } from '../canvas-strategies'
@@ -10,12 +21,17 @@ import {
   getTargetPathsFromInteractionTarget,
   strategyApplicationResult,
 } from '../canvas-strategy-types'
-import type { InteractionSession } from '../interaction-state'
+import type { DragInteractionData, InteractionSession } from '../interaction-state'
+import { getClosestGridCellToPoint, gridCellCoordinates } from './grid-cell-bounds'
+import type { GridCellGlobalFrames, SortableGridElementProperties } from './grid-helpers'
 import {
-  getCommandsAndPatchForGridRearrange,
+  findOriginalGrid,
+  getOriginalElementGridConfiguration,
   getParentGridTemplatesFromChildMeasurements,
   gridMoveStrategiesExtraCommands,
-} from './grid-move-helpers'
+  setGridPropsCommands,
+  sortElementsByGridPosition,
+} from './grid-helpers'
 
 export const gridMoveRearrangeStrategy: CanvasStrategyFactory = (
   canvasState: InteractionCanvasState,
@@ -104,4 +120,151 @@ export const gridMoveRearrangeStrategy: CanvasStrategyFactory = (
       )
     },
   }
+}
+
+function getCommandsAndPatchForGridRearrange(
+  canvasState: InteractionCanvasState,
+  interactionData: DragInteractionData,
+  selectedElement: ElementPath,
+): {
+  commands: CanvasCommand[]
+  elementsToRerender: ElementPath[]
+} {
+  if (interactionData.drag == null) {
+    return { commands: [], elementsToRerender: [] }
+  }
+
+  const selectedElementMetadata = MetadataUtils.findElementByElementPath(
+    canvasState.startingMetadata,
+    selectedElement,
+  )
+  if (selectedElementMetadata == null) {
+    return { commands: [], elementsToRerender: [] }
+  }
+
+  const gridPath = findOriginalGrid(canvasState.startingMetadata, EP.parentPath(selectedElement)) // TODO don't use EP.parentPath
+  if (gridPath == null) {
+    return { commands: [], elementsToRerender: [] }
+  }
+
+  const { parentGridCellGlobalFrames, parentContainerGridProperties } =
+    selectedElementMetadata.specialSizeMeasurements
+  if (parentGridCellGlobalFrames == null) {
+    return { commands: [], elementsToRerender: [] }
+  }
+
+  const commands = runGridMoveRearrange(
+    canvasState.startingMetadata,
+    interactionData,
+    selectedElementMetadata,
+    gridPath,
+    parentGridCellGlobalFrames,
+    parentContainerGridProperties,
+    null,
+  )
+
+  return {
+    commands: commands,
+    elementsToRerender: [EP.parentPath(selectedElement), selectedElement],
+  }
+}
+
+export function runGridMoveRearrange(
+  jsxMetadata: ElementInstanceMetadataMap,
+  interactionData: DragInteractionData,
+  selectedElementMetadata: ElementInstanceMetadata,
+  gridPath: ElementPath,
+  gridCellGlobalFrames: GridCellGlobalFrames,
+  gridTemplate: GridContainerProperties,
+  newPathAfterReparent: ElementPath | null,
+): CanvasCommand[] {
+  if (interactionData.drag == null) {
+    return []
+  }
+
+  const isReparent = newPathAfterReparent != null
+  const pathForCommands = isReparent ? newPathAfterReparent : selectedElementMetadata.elementPath // when reparenting, we want to use the new path for commands
+
+  const gridConfig = isReparent
+    ? {
+        originalCellBounds: { width: 1, height: 1 }, // when reparenting, we just put it in a single cell
+        mouseCellPosInOriginalElement: { row: 0, column: 0 },
+      }
+    : getOriginalElementGridConfiguration(
+        gridCellGlobalFrames,
+        interactionData,
+        selectedElementMetadata,
+      )
+  if (gridConfig == null) {
+    return []
+  }
+  const { mouseCellPosInOriginalElement, originalCellBounds } = gridConfig
+
+  const mousePos = offsetPoint(interactionData.dragStart, interactionData.drag)
+  const targetCellData = getClosestGridCellToPoint(gridCellGlobalFrames, mousePos)
+  const targetCellCoords = targetCellData?.gridCellCoordinates
+  if (targetCellCoords == null) {
+    return []
+  }
+
+  const row = Math.max(targetCellCoords.row - mouseCellPosInOriginalElement.row, 1)
+  const column = Math.max(targetCellCoords.column - mouseCellPosInOriginalElement.column, 1)
+
+  const gridCellMoveCommands = setGridPropsCommands(pathForCommands, gridTemplate, {
+    gridColumnStart: gridPositionValue(column),
+    gridColumnEnd: gridPositionValue(column + originalCellBounds.height),
+    gridRowStart: gridPositionValue(row),
+    gridRowEnd: gridPositionValue(row + originalCellBounds.width),
+  })
+
+  // The siblings of the grid element being moved
+  const siblings = MetadataUtils.getChildrenUnordered(jsxMetadata, gridPath)
+    .filter((s) => !EP.pathsEqual(s.elementPath, selectedElementMetadata.elementPath))
+    .map(
+      (s, index): SortableGridElementProperties => ({
+        ...s.specialSizeMeasurements.elementGridProperties,
+        index: index,
+        path: s.elementPath,
+      }),
+    )
+
+  // Sort the siblings and the cell under mouse ascending based on their grid coordinates, so that
+  // the indexes grow left-right, top-bottom.
+  const templateColumnsCount =
+    gridTemplate.gridTemplateColumns?.type === 'DIMENSIONS'
+      ? gridTemplate.gridTemplateColumns.dimensions.length
+      : 1
+  const cellsSortedByPosition = siblings
+    .concat({
+      ...{
+        gridColumnStart: gridPositionValue(targetCellCoords.column),
+        gridColumnEnd: gridPositionValue(targetCellCoords.column),
+        gridRowStart: gridPositionValue(targetCellCoords.row),
+        gridRowEnd: gridPositionValue(targetCellCoords.row),
+      },
+      path: selectedElementMetadata.elementPath,
+      index: siblings.length + 1,
+    })
+    .sort(sortElementsByGridPosition(templateColumnsCount))
+
+  const indexInSortedCellsForRearrange = cellsSortedByPosition.findIndex((s) =>
+    EP.pathsEqual(selectedElementMetadata.elementPath, s.path),
+  )
+
+  const updateGridControlsCommand = showGridControls(
+    'mid-interaction',
+    gridPath,
+    targetCellData?.gridCellCoordinates ?? null,
+    gridCellCoordinates(row, column),
+  )
+
+  return [
+    ...gridCellMoveCommands,
+    reorderElement(
+      'always',
+      pathForCommands,
+      absolute(Math.max(indexInSortedCellsForRearrange, 0)),
+    ),
+    updateGridControlsCommand,
+  ]
 }
